--- packages/util/src/index.ts
@@ -1,3 +1,0 @@
-export function testIndex(): void {
-  console.log("testIndex");
-}
